@@ -12,31 +12,33 @@
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
-    "@remix-run/node": "^2.12.1",
-    "@remix-run/react": "^2.12.1",
-    "@remix-run/serve": "^2.12.1",
+    "@dotenvx/dotenvx": "^1.14.2",
+    "@remix-run/node": "^2.15.0",
+    "@remix-run/react": "^2.15.0",
+    "@remix-run/serve": "^2.15.0",
+    "@valley/db": "*",
     "@valley/shared": "*",
     "@valley/ui": "*",
-    "@valley/db": "*",
+    "@vercel/remix": "2.15.0",
     "classnames": "^2.5.1",
     "isbot": "^4.1.0",
     "overlayscrollbars": "^2.10.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "chalk": "^5.3.0"
+    "react": "^18.3.1",
+    "ansis": "^3.3.2",
+    "react-dom": "^18.3.1"
   },
   "devDependencies": {
-    "cross-env": "^7.0.3",
     "@csstools/postcss-media-minmax": "^2.0.1",
-    "@remix-run/dev": "^2.12.1",
-    "@types/react": "^18.2.20",
-    "@types/react-dom": "^18.2.7",
+    "@remix-run/dev": "^2.15.0",
+    "@types/react": "^18.3.12",
+    "@types/react-dom": "^18.3.1",
     "@valley/eslint-config": "*",
     "@valley/typescript-config": "*",
     "autoprefixer": "^10.4.19",
+    "cross-env": "^7.0.3",
     "eslint": "^8.56.11",
     "postcss": "^8.4.38",
-    "typescript": "^5.3.3",
+    "typescript": "^5.7.2",
     "vite": "^5.1.0",
     "vite-tsconfig-paths": "^4.2.1"
   },
