@@ -1,69 +1,146 @@
 {
   "name": "web",
-  "version": "0.0.0",
+  "version": "0.0.1",
   "private": true,
+  "sideEffects": false,
+  "type": "module",
   "scripts": {
-    "dev": "yarn fix-vips-concurrency-env && next dev --turbo --port \"$(dotenvx get WEB_PORT -f ../../.env)\"",
-    "build": "next build",
-    "start": "yarn fix-vips-concurrency-env && next start",
+    "build": "cross-env NODE_ENV=production remix vite:build && tsc --project ./tsconfig.server.json",
+    "dev": "cross-env NODE_ENV=development tsx watch --clear-screen=false --ignore \".cache/**\" --ignore \"app/**\" --ignore \"vite.config.ts.timestamp-*\" --ignore \"build/**\" --ignore \"node_modules/**\" --inspect ./server/index.ts",
+    "remix:dev": "yarn dotenv remix vite:dev",
+    "start": "cross-env NODE_ENV=production node ./build/runner/index.js",
+    "lint": "eslint . --max-warnings 0 --ignore-path ../../.gitignore --cache --cache-location ./node_modules/.cache/eslint",
     "typecheck": "tsc --noEmit",
-    "lint": "eslint . --max-warnings 0",
-    "fix-vips-concurrency-env": "VIPS_CONCURRENCY=$(nproc 2>/dev/null || sysctl -n hw.logicalcpu)"
+    "deploy:cf": "npx wrangler pages deploy ./build/client",
+    "dotenv": "dotenvx run -f ../../.env --"
   },
   "dependencies": {
+    "@conform-to/react": "^1.2.2",
+    "@conform-to/zod": "^1.2.2",
+    "@dnd-kit/core": "^6.3.1",
+    "@dnd-kit/modifiers": "^9.0.0",
+    "@dnd-kit/sortable": "^10.0.0",
     "@dotenvx/dotenvx": "^1.14.2",
-    "@storybook/react": "^8.2.7",
-    "@uppy/core": "^4.2.0",
-    "@uppy/dashboard": "^4.1.0",
-    "@uppy/drag-drop": "^4.0.2",
-    "@uppy/file-input": "^4.0.1",
+    "@epic-web/cachified": "^5.2.0",
+    "@epic-web/client-hints": "^1.3.5",
+    "@epic-web/totp": "^2.0.0",
+    "@faker-js/faker": "^9.1.0",
+    "@hookform/resolvers": "^3.9.1",
+    "@nasa-gcn/remix-seo": "^2.0.1",
+    "@neondatabase/serverless": "^0.10.4",
+    "@paralleldrive/cuid2": "^2.2.2",
+    "@prisma/adapter-neon": "^6.1.0",
+    "@prisma/adapter-pg": "^6.0.1",
+    "@react-email/components": "^0.0.25",
+    "@react-router/dev": "^0.0.0-nightly-bf7ecb711-20240911",
+    "@remix-pwa/dev": "^3.1.0",
+    "@remix-pwa/sw": "^3.0.10",
+    "@remix-pwa/worker-runtime": "^2.1.4",
+    "@remix-run/cloudflare": "^2.15.2",
+    "@remix-run/cloudflare-pages": "^2.15.2",
+    "@remix-run/express": "^2.15.2",
+    "@remix-run/node": "^2.15.2",
+    "@remix-run/react": "^2.15.2",
+    "@remix-run/serve": "^2.15.2",
+    "@remix-run/server-runtime": "^2.15.2",
+    "@sentry/profiling-node": "^8.36.0",
+    "@sentry/remix": "^8.36.0",
+    "@uppy/core": "^4.1.0",
     "@uppy/progress-bar": "^4.0.0",
-    "@uppy/react": "^4.0.2",
-    "@uppy/status-bar": "^4.0.3",
-    "@uppy/tus": "^4.1.1",
+    "@uppy/tus": "^4.1.4",
+    "@valley/db": "*",
     "@valley/shared": "*",
     "@valley/ui": "*",
-    "axios": "^1.7.3",
-    "axios-retry": "^4.5.0",
+    "add": "^2.0.6",
+    "address": "^2.0.3",
+    "ansis": "^3.3.2",
+    "axios": "^1.7.7",
+    "bcryptjs": "^2.4.3",
+    "cachified-redis-json-adapter": "^2.0.0",
     "classnames": "^2.5.1",
-    "dayjs": "^1.11.12",
-    "framer-motion": "^11.3.30",
-    "geist": "^1.3.1",
+    "close-with-grace": "^2.1.0",
+    "compression": "^1.7.4",
+    "cookie": "^1.0.1",
+    "cross-env": "^7.0.3",
+    "crypto-js": "^4.2.0",
+    "dayjs": "^1.11.13",
+    "express": "^4.21.1",
+    "express-rate-limit": "^7.4.1",
+    "framer-motion": "^11.11.10",
     "geist-ui-icons": "^1.1.1",
+    "get-port": "^7.1.0",
+    "helmet": "^8.0.0",
     "immer": "^10.1.1",
-    "jwt-decode": "^4.0.0",
+    "input-otp": "^1.2.4",
+    "intl-parse-accept-language": "^1.0.0",
+    "is-ip": "^5.0.1",
+    "isbot": "^4.1.0",
+    "libphonenumber-js": "^1.11.13",
+    "localforage": "^1.10.0",
     "lodash.deburr": "^4.1.0",
-    "next": "^14.2.7",
-    "next-auth": "^4.24.7",
+    "lru-cache": "^11.0.1",
+    "memfs": "^4.14.0",
+    "morgan": "^1.10.0",
+    "msw": "^2.6.0",
     "overlayscrollbars": "^2.10.0",
     "overlayscrollbars-react": "^0.5.6",
     "pretty-bytes": "^6.1.1",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-hook-form": "^7.52.2",
-    "react-is": "^18.3.1",
-    "react-textarea-autosize": "^8.5.3",
-    "sharp": "^0.33.5",
-    "swr": "^2.2.5",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-hook-form": "^7.53.1",
+    "react-phone-number-input": "^3.4.9",
+    "react-router": "^0.0.0-nightly-f47048541-20240813",
+    "redis": "^4.7.0",
+    "remix-auth": "^3.7.0",
+    "remix-auth-github": "^2.0.0",
+    "remix-auth-totp": "^3.4.2",
+    "remix-auth-vk": "^1.1.3",
+    "remix-auth-webauthn": "^0.3.4",
+    "remix-client-cache": "^1.1.0",
+    "remix-hook-form": "^5.1.1",
+    "remix-utils": "^7.7.0",
+    "sonner": "^1.7.0",
+    "source-map-support": "^0.5.21",
+    "transliteration": "^2.3.5",
+    "turbo-stream": "^2.4.1",
     "tus-js-client": "^4.2.3",
-    "uuid": "^10.0.0",
-    "vaul": "^0.9.6",
-    "zustand": "^4.5.5"
+    "usehooks-ts": "^3.1.0",
+    "vite-env-only": "^3.0.3",
+    "vite-plugin-node-polyfills": "^0.22.0",
+    "web-oidc": "^0.0.19",
+    "ws": "^8.18.0",
+    "zod": "^3.23.8",
+    "zustand": "^5.0.1"
   },
   "devDependencies": {
+    "@cloudflare/workers-types": "^4.20241127.0",
     "@csstools/postcss-media-minmax": "^2.0.1",
-    "@next/eslint-plugin-next": "^14.2.7",
-    "@swc/helpers": "^0.5.12",
+    "@hookform/devtools": "^4.3.1",
+    "@react-router/remix-config-routes-adapter": "^0.0.0-nightly-bf7ecb711-20240911",
+    "@remix-run/dev": "^2.15.2",
+    "@types/bcryptjs": "^2.4.6",
+    "@types/compression": "^1.7.5",
+    "@types/express": "^5.0.0",
     "@types/lodash.deburr": "^4.1.9",
-    "@types/node": "^20.11.24",
-    "@types/react": "^18.2.61",
-    "@types/react-dom": "^18.2.19",
-    "@types/react-is": "^18.3.0",
-    "@valley/db": "*",
+    "@types/morgan": "^1.9.9",
+    "@types/react": "^18.3.12",
+    "@types/react-dom": "^18.3.1",
+    "@types/source-map-support": "^0.5.10",
+    "@types/ws": "^8.5.13",
     "@valley/eslint-config": "*",
     "@valley/typescript-config": "*",
+    "@vercel/remix": "2.15.0",
+    "autoprefixer": "^10.4.19",
     "eslint": "^8.56.11",
-    "typescript": "^5.3.3",
-    "typescript-plugin-css-modules": "^5.1.0"
+    "postcss": "^8.4.38",
+    "react-email": "^3.0.1",
+    "remix-flat-routes": "^0.6.5",
+    "tsx": "^4.19.2",
+    "typescript": "^5.7.2",
+    "vite": "^5.1.0",
+    "vite-tsconfig-paths": "^4.2.1"
+  },
+  "engines": {
+    "node": ">=20.0.0"
   }
 }
