@@ -5,8 +5,6 @@
   "files": [
     "base.js",
     "library.js",
-    "nest.js",
-    "next.js",
     "remix.js",
     "prettier-base.js",
     "react-internal.js"
@@ -23,6 +21,6 @@
     "eslint-plugin-prettier": "^5.2.1",
     "eslint-plugin-react": "^7.37.1",
     "eslint-plugin-react-hooks": "^4.6.2",
-    "typescript": "^5.3.3"
+    "typescript": "^5.7.2"
   }
 }
