@@ -3,22 +3,29 @@
   "version": "0.0.0",
   "private": true,
   "exports": {
+    "./Skeleton": "./src/Skeleton/Skeleton.tsx",
+    "./Portal": "./src/Portal/Portal.tsx",
+    "./Menu": "./src/Menu/Menu.tsx",
+    "./ContextMenu": "./src/ContextMenu/ContextMenu.tsx",
+    "./Popover": "./src/Popover/Popover.tsx",
+    "./AnimateChangeInHeight": "./src/AnimateChangeInHeight/AnimateChangeInHeight.tsx",
+    "./Switch": "./src/Switch/Switch.tsx",
+    "./Animated": "./src/Animated/Animated.tsx",
+    "./FormHelperText": "./src/FormHelperText/FormHelperText.tsx",
+    "./FormControl": "./src/FormControl/index.ts",
+    "./TextField": "./src/TextField/TextField.tsx",
+    "./OTPInput": "./src/OTPInput/OTPInput.tsx",
+    "./Hidden": "./src/Hidden/Hidden.tsx",
     "./Stack": "./src/Stack/Stack.tsx",
     "./Note": "./src/Note/Note.tsx",
-    "./MenuSeparator": "./src/MenuSeparator/MenuSeparator.tsx",
     "./Gauge": "./src/Gauge/Gauge.tsx",
     "./usePageScroll": "./src/usePageScroll/usePageScroll.ts",
     "./useScrollProgress": "./src/useScrollProgress/useScrollProgress.ts",
-    "./useContextMenu": "./src/useContextMenu/useContextMenu.ts",
-    "./MenuIconButton": "./src/MenuIconButton/MenuIconButton.tsx",
-    "./MenuButton": "./src/MenuButton/MenuButton.tsx",
-    "./Dropdown": "./src/Dropdown/Dropdown.tsx",
+    "./DropdownMenu": "./src/DropdownMenu/DropdownMenu.tsx",
     "./Grow": "./src/Grow/Grow.tsx",
     "./Fade": "./src/Fade/Fade.tsx",
     "./Paper": "./src/Paper/Paper.tsx",
     "./TextArea": "./src/TextArea/TextArea.tsx",
-    "./MenuItem": "./src/MenuItem/MenuItem.tsx",
-    "./Menu": "./src/Menu/Menu.tsx",
     "./Divider": "./src/Divider/Divider.tsx",
     "./Wrapper": "./src/Wrapper/Wrapper.tsx",
     "./ModalHeader": "./src/ModalHeader/ModalHeader.tsx",
@@ -37,7 +44,8 @@
     "./useMediaQuery": "./src/useMediaQuery/useMediaQuery.ts",
     "./config/theme": "./src/config/theme.ts",
     "./styles/global.css": "./src/styles/global.css",
-    "./styles/theme.css": "./src/styles/theme.css"
+    "./styles/theme.css": "./src/styles/theme.css",
+    "./styles/reset.css": "./src/styles/reset.css"
   },
   "scripts": {
     "typecheck": "tsc --noEmit",
@@ -48,20 +56,36 @@
     "@turbo/gen": "^1.12.4",
     "@types/eslint": "^8.56.11",
     "@types/node": "^20.11.24",
-    "@types/react": "^18.2.61",
-    "@types/react-dom": "^18.2.19",
+    "@types/react": "^18.3.12",
+    "@types/react-dom": "^18.3.1",
+    "@types/react-is": "^18.3.0",
     "@valley/eslint-config": "*",
-    "@valley/shared": "*",
     "@valley/typescript-config": "*",
     "eslint": "^8.56.11",
-    "react": "^18.2.0",
-    "typescript": "^5.3.3",
+    "typescript": "^5.7.2",
     "typescript-plugin-css-modules": "^5.1.0"
   },
   "dependencies": {
-    "@mui/base": "^5.0.0-beta.58",
-    "@radix-ui/react-slot": "^1.1.0",
+    "@floating-ui/react-dom": "^2.1.2",
+    "@formkit/auto-animate": "^0.8.2",
+    "@radix-ui/react-context-menu": "^2.2.4",
+    "@radix-ui/react-dialog": "^1.1.4",
+    "@radix-ui/react-dropdown-menu": "^2.1.4",
+    "@radix-ui/react-popover": "^1.1.4",
+    "@radix-ui/react-portal": "^1.1.3",
+    "@radix-ui/react-slot": "^1.1.1",
+    "@storybook/react": "^8.4.1",
+    "@valley/shared": "*",
     "classnames": "^2.5.1",
-    "geist-ui-icons": "^1.1.1"
+    "framer-motion": "^11.11.10",
+    "geist-ui-icons": "^1.1.1",
+    "input-otp": "^1.2.4",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-hook-form": "^7.53.1",
+    "react-is": "^18.3.1",
+    "react-router-dom": "^6.27.0",
+    "react-textarea-autosize": "^8.5.4",
+    "vaul": "^1.1.2"
   }
 }
