@@ -1,6 +1,6 @@
 import { remember } from '../utils/remember'
-import { PrismaClient } from '@prisma/client'
-import chalk from 'chalk'
+import { PrismaClient } from '@valley/db'
+import ansis from 'ansis'
 
 // NOTE: if you change anything in this function you'll need to restart
 // the dev server to see your changes.
@@ -26,7 +26,7 @@ export const prisma = remember('prisma', () => {
             : e.duration < logThreshold * 1.4
               ? 'redBright'
               : 'red'
-    const dur = chalk[color](`${e.duration}ms`)
+    const dur = ansis[color](`${e.duration}ms`)
     console.info(`prisma:query - ${dur} - ${e.query}`)
   })
 
