@@ -1,4 +1,4 @@
-import type { StorybookConfig } from '@storybook/nextjs'
+import type { StorybookConfig } from '@storybook/react-vite'
 import { join, dirname } from 'path'
 
 function getAbsolutePath(value: string): any {
@@ -6,16 +6,13 @@ function getAbsolutePath(value: string): any {
 }
 
 const config: StorybookConfig = {
-  stories: ['../../../packages/ui/src/**/*.stories.@(js|jsx|ts|tsx|mdx)'],
+  stories: ['../../../packages/ui/src/**/*.@(mdx|stories.@(js|jsx|ts|tsx))'],
+  staticDirs: ['../public'],
   addons: [getAbsolutePath('@storybook/addon-essentials')],
-  framework: {
-    name: getAbsolutePath('@storybook/nextjs'),
-    options: {
-      nextConfigPath: '../next.config.mjs',
-      builder: {
-        useSWC: true,
-      },
-    },
+  framework: getAbsolutePath('@storybook/react-vite'),
+  docs: {},
+  typescript: {
+    reactDocgen: 'react-docgen-typescript',
   },
 }
 
