@@ -1,10 +1,7 @@
 import React from 'react'
 import ButtonBase from '@valley/ui/ButtonBase'
 import styles from './FolderCard.module.css'
-import { formatBytes } from '../../utils/formatBytes'
 import cx from 'classnames'
-import { SerializedFolder } from '@valley/shared'
-import MenuIconButton from '@valley/ui/MenuIconButton'
 import {
   MoreVertical,
   PencilEdit,
@@ -12,54 +9,101 @@ import {
   Menu as MenuIcon,
 } from 'geist-ui-icons'
 import Menu from '@valley/ui/Menu'
-import Dropdown from '@valley/ui/Dropdown'
-import MenuItem from '@valley/ui/MenuItem'
-import { useContextMenu } from '@valley/ui/useContextMenu'
-import MenuSeparator from '@valley/ui/MenuSeparator'
-import { useParams, useRouter, useSearchParams } from 'next/navigation'
+import { formatBytes } from 'app/utils/misc'
+import { Folder } from '@valley/db'
+import IconButton from '@valley/ui/IconButton'
+import { Link, useParams, useSearchParams } from '@remix-run/react'
 
 type FolderCardProps = {
-  folder: SerializedFolder
-  active?: boolean
-  onClick?: (folder: SerializedFolder) => void
+  folder: Folder
+  onClick?: (folder: Folder) => void
 }
 
-const FolderCard: React.FC<FolderCardProps> = ({ folder, active, onClick }) => {
-  const totalSize = formatBytes(folder.totalSize)
-  const { dropdownProps, onContextMenu, onMenuButtonClick, anchor } =
-    useContextMenu()
-  const { id } = useParams()
-  const searchParams = useSearchParams()
-  const router = useRouter()
-
-  const handleClick = () => {
-    onClick?.(folder)
-  }
+const FolderCardMenuContent: React.FC<{
+  folder: Folder
+}> = ({ folder }) => {
+  const [searchParams, setSearchParams] = useSearchParams()
 
   const handleFolderRename = () => {
     const params = new URLSearchParams(searchParams.toString())
     params.set('modal', 'edit-folder-title')
     params.set('modal-folderId', folder.id.toString())
-    router.push(`/projects/${id}?${params.toString()}`)
+    setSearchParams(params, {
+      state: {
+        defaultTitle: folder.title,
+      },
+      preventScrollReset: true,
+    })
   }
 
   const handleFolderDelete = () => {
-    const params = new URLSearchParams(searchParams.toString())
-    params.set('modal', 'confirm-folder-deletion')
-    params.set('modal-folderId', folder.id.toString())
-    router.push(`/projects/${id}?${params.toString()}`)
+    searchParams.set('modal', 'confirm-folder-deletion')
+    searchParams.set('modal-folderId', folder.id.toString())
+    setSearchParams(searchParams, {
+      preventScrollReset: true,
+    })
+  }
+
+  const handleFolderClear = () => {
+    searchParams.set('modal', 'confirm-folder-clear')
+    searchParams.set('modal-folderId', folder.id.toString())
+    setSearchParams(searchParams, {
+      preventScrollReset: true,
+    })
   }
 
   return (
-    <Dropdown {...dropdownProps}>
+    <Menu.Content>
+      <Menu.Item
+        onClick={handleFolderRename}
+        before={<PencilEdit color="var(--text-secondary)" />}
+      >
+        Rename
+      </Menu.Item>
+      <Menu.Item before={<MenuIcon color="var(--text-secondary)" />}>
+        Reorder
+      </Menu.Item>
+      <Menu.Separator />
+      {folder.isDefaultFolder && (
+        <>
+          <Menu.Item
+            onClick={handleFolderClear}
+            before={<Trash color="var(--red-600)" />}
+          >
+            Delete all files
+          </Menu.Item>
+        </>
+      )}
+      {!folder.isDefaultFolder && (
+        <Menu.Item
+          onClick={handleFolderDelete}
+          before={<Trash color="var(--red-600)" />}
+        >
+          Delete
+        </Menu.Item>
+      )}
+    </Menu.Content>
+  )
+}
+
+const FolderCard: React.FC<FolderCardProps> = ({ folder, onClick }) => {
+  const { projectId, folderId } = useParams()
+  const isActive = folderId ? folderId === folder.id : folder.isDefaultFolder
+  const totalSize = formatBytes(Number(folder.totalSize))
+
+  return (
+    <Menu.Root openOnContextMenu>
       <ButtonBase
         asChild
         variant="secondary"
         className={cx(styles.folderCard, {
-          [styles['folderCard--active']]: active,
+          [styles['folderCard--active']]: isActive,
         })}
       >
-        <a onClick={handleClick} onContextMenu={onContextMenu}>
+        <Link
+          onClick={onClick?.bind(null, folder)}
+          to={'/projects/' + projectId + '/folder/' + folder.id}
+        >
           <div className={styles.folderCard__content}>
             <h5 className={styles.folderCard__contentTitle}>{folder.title}</h5>
             <div className={styles.folderCard__contentSubtitle}>
@@ -69,39 +113,23 @@ const FolderCard: React.FC<FolderCardProps> = ({ folder, active, onClick }) => {
             </div>
           </div>
 
-          <MenuIconButton
-            onClick={onMenuButtonClick}
-            size="sm"
-            variant="tertiary"
-          >
-            <MoreVertical />
-          </MenuIconButton>
-          <Menu
-            container={window.document.body}
-            id={'folderMenu-' + folder.id.toString()}
-            placement="bottom-start"
-            anchor={anchor}
-          >
-            <MenuItem
-              onClick={handleFolderRename}
-              before={<PencilEdit color="var(--text-secondary)" />}
+          <Menu.Trigger asChild>
+            <IconButton
+              onClick={(e) => {
+                e.stopPropagation()
+                e.preventDefault()
+              }}
+              size="sm"
+              variant="tertiary"
             >
-              Rename
-            </MenuItem>
-            <MenuItem before={<MenuIcon color="var(--text-secondary)" />}>
-              Reorder
-            </MenuItem>
-            <MenuSeparator />
-            <MenuItem
-              onClick={handleFolderDelete}
-              before={<Trash color="var(--red-600)" />}
-            >
-              Delete
-            </MenuItem>
-          </Menu>
-        </a>
+              <MoreVertical />
+            </IconButton>
+          </Menu.Trigger>
+
+          <FolderCardMenuContent folder={folder} />
+        </Link>
       </ButtonBase>
-    </Dropdown>
+    </Menu.Root>
   )
 }
 
