@@ -1,86 +1,89 @@
-'use client'
-import React, { useState } from 'react'
+import React from 'react'
 import Button from '@valley/ui/Button'
-import Input from '@valley/ui/Input'
-import InputLabel from '@valley/ui/InputLabel'
 import ModalHeader from '@valley/ui/ModalHeader'
 import ModalFooter from '@valley/ui/ModalFooter'
-import styles from './CreateProject.module.css'
-import { SubmitHandler, useForm } from 'react-hook-form'
-import { createProject } from '../../../api/projects'
-import { useRouter } from 'next/navigation'
+import Modal from '@valley/ui/Modal'
+import styles from '../Modals.module.css'
+import { useRemixForm } from 'remix-hook-form'
+import z from 'zod'
+import { zodResolver } from '@hookform/resolvers/zod'
+import TextField from '@valley/ui/TextField'
+import { Form, useFetcher } from '@remix-run/react'
+import {
+  ProjectsCreateSchema,
+  type action as createAction,
+} from 'app/routes/api+/projects+/create'
+
+type FormData = z.infer<typeof ProjectsCreateSchema>
+
+const resolver = zodResolver(ProjectsCreateSchema)
 
 type CreateProjectModalProps = {
   onClose: () => void
 }
 
-type FieldValues = {
-  projectName: string
-  dateShot: Date
-}
-
 const CreateProjectModal: React.FC<CreateProjectModalProps> = ({ onClose }) => {
-  const [isLoading, setIsLoading] = useState(false)
-  const router = useRouter()
-  const { register, handleSubmit } = useForm<FieldValues>()
-
-  const onSubmit: SubmitHandler<FieldValues> = async (values, e) => {
-    e?.preventDefault()
-
-    setIsLoading(true)
-    const res = await createProject({
-      dateShot: values.dateShot,
-      protected: false,
-      title: values.projectName,
-      storedUntil: new Date(),
-      translationStringsId: null,
-    })
-
-    onClose()
-    router.replace('/projects/' + res.project.id)
-  }
+  const fetcher = useFetcher<typeof createAction>({
+    key: 'projects-create',
+  })
+  const { register, handleSubmit } = useRemixForm<FormData>({
+    resolver,
+    submitConfig: {
+      action: '/api/projects/create',
+      method: 'POST',
+    },
+    fetcher,
+  })
+  const isPending = fetcher.state !== 'idle'
 
   return (
-    <>
+    <Modal onDismiss={onClose} id="create-project">
       <ModalHeader>Create Project</ModalHeader>
-      <form
+      <Form
+        onSubmit={handleSubmit}
+        className={styles.modal__content}
         id="create-project-form"
-        className={styles.createProjectModal__form}
-        onSubmit={handleSubmit(onSubmit)}
+        method="POST"
+        action="/api/projects/create"
       >
-        <div>
-          <InputLabel htmlFor="project-name-input">Project Name *</InputLabel>
-          <Input
-            {...register('projectName')}
-            id="project-name-input"
-            placeholder="my-project"
-          />
-        </div>
-        <div>
-          <InputLabel htmlFor="store-until-select">Store until *</InputLabel>
-          <Input id="store-until-select" placeholder="my-project" />
-        </div>
-        <div>
-          <InputLabel htmlFor="visibility-select">Visibility *</InputLabel>
-          <Input id="visibility-select" placeholder="my-project" />
-        </div>
-        <div>
-          <InputLabel htmlFor="date-shot-input">Date shot</InputLabel>
-          <Input
-            {...register('dateShot', {
-              valueAsDate: true,
-            })}
-            id="date-shot-input"
-            type="date"
-            placeholder="dd.mm.yyyy"
-          />
-        </div>
-      </form>
+        <TextField
+          {...register('projectName')}
+          label="Project Name"
+          required
+          size="lg"
+          id="project-name-input"
+          placeholder="my-project"
+        />
+        <TextField
+          label="Store until"
+          required
+          size="lg"
+          id="store-until-select"
+          placeholder="my-project"
+        />
+        <TextField
+          label="Visibility"
+          required
+          size="lg"
+          id="visibility-select"
+          placeholder="my-project"
+        />
+        <TextField
+          {...register('dateShot', {
+            valueAsDate: true,
+          })}
+          label="Date shot"
+          size="lg"
+          id="date-shot-input"
+          type="date"
+          placeholder="dd.mm.yyyy"
+        />
+      </Form>
       <ModalFooter
         before={
           <Button
             onClick={onClose}
-            disabled={isLoading}
+            disabled={isPending}
             variant="secondary-dimmed"
             size="md"
           >
@@ -93,14 +96,14 @@ const CreateProjectModal: React.FC<CreateProjectModalProps> = ({ onClose }) => {
             variant="primary"
             size="md"
             type="submit"
-            disabled={isLoading}
-            loading={isLoading}
+            disabled={isPending}
+            loading={isPending}
           >
             Create
           </Button>
         }
       />
-    </>
+    </Modal>
   )
 }
 
