@@ -1,103 +1,82 @@
-'use client'
-import React, { useEffect, useState } from 'react'
+import React, { Suspense } from 'react'
 import Button from '@valley/ui/Button'
-import Input from '@valley/ui/Input'
-import InputLabel from '@valley/ui/InputLabel'
 import ModalHeader from '@valley/ui/ModalHeader'
 import ModalFooter from '@valley/ui/ModalFooter'
-import styles from './EditFolderTitle.module.css'
-import { SubmitHandler, useForm } from 'react-hook-form'
-import { editFolder } from '../../../api/folders'
-import { useParams, useSearchParams } from 'next/navigation'
-import useSWR from 'swr'
-import {
-  PROJECT_FOLDER_TITLE_MAX_LENGTH,
-  ProjectGetReq,
-  ProjectGetRes,
-} from '@valley/shared'
-import { api } from '../../../api'
+import styles from '../Modals.module.css'
+import z from 'zod'
+import { zodResolver } from '@hookform/resolvers/zod'
+import TextField from '@valley/ui/TextField'
+import { Await, Form, useParams, useSearchParams } from '@remix-run/react'
+import { FoldersEditSchema } from 'app/routes/api+/folders+/$id.edit'
+import { useRemixForm } from 'remix-hook-form'
+import { useIsPending } from 'app/utils/misc'
+import Modal from '@valley/ui/Modal'
+import { useProjectAwait } from 'app/utils/project'
+import { ProjectWithFolders } from '@valley/shared'
+
+type FormData = z.infer<typeof FoldersEditSchema>
+
+const resolver = zodResolver(FoldersEditSchema)
 
 type EditFolderTitleModalProps = {
   onClose: () => void
 }
 
-type FieldValues = {
-  folderTitle: string
-}
-
-const EditFolderTitleModal: React.FC<EditFolderTitleModalProps> = ({
-  onClose,
-}) => {
-  const [isLoading, setIsLoading] = useState(false)
-  const { id: projectId } = useParams()
-  const { data, mutate } = useSWR<ProjectGetRes, ProjectGetReq>(
-    '/projects/' + projectId,
-    api({ isAccessTokenRequired: true })
-  )
-  const searchParams = useSearchParams()
-  const modalPropsFolderId = searchParams.get('modal-folderId')
-  const searchParamsFolderId = searchParams.get('folder')
-  const folderId = Number(modalPropsFolderId || searchParamsFolderId)
-  const parsedProjectId = Number(projectId)
-  const defaultTitle = data?.folders.find((e) => e.id === folderId)?.title
-  const { register, handleSubmit, setFocus } = useForm<FieldValues>()
-
-  const onSubmit: SubmitHandler<FieldValues> = async (values, e) => {
-    e?.preventDefault()
-    if (isNaN(parsedProjectId) || isNaN(folderId) || !data) return
-
-    setIsLoading(true)
-    const res = await editFolder(
-      {
-        title: values.folderTitle,
-        id: folderId,
-      },
-      parsedProjectId
-    )
-
-    onClose()
-    await mutate({
-      ...data,
-      folders: data.folders.map((e) =>
-        e.id === res.folder.id ? res.folder : e
-      ),
-    })
+const ModalContent: React.FC<
+  EditFolderTitleModalProps & {
+    project?: ProjectWithFolders | null
   }
-
-  useEffect(() => {
-    setFocus('folderTitle', {
-      shouldSelect: true,
+> = ({ onClose, project }) => {
+  const { folderId: paramsFolderId } = useParams()
+  const [searchParams] = useSearchParams()
+  const modalPropsFolderId = searchParams.get('modal-folderId')
+  const folderId = modalPropsFolderId || paramsFolderId
+  const currentFolder = project?.folders.find((f) => f.id === folderId)
+  const defaultTitle = currentFolder?.title
+  const formAction = '/api/folders/' + folderId + '/edit'
+  const { register, getFieldState, formState, handleSubmit } =
+    useRemixForm<FormData>({
+      resolver,
+      submitConfig: {
+        action: formAction,
+        method: 'POST',
+      },
     })
-  }, [])
+  const isPending = useIsPending({
+    formMethod: 'POST',
+    formAction,
+  })
 
   return (
-    <>
+    <Modal onDismiss={onClose} id="edit-folder-title">
       <ModalHeader>Edit Folder Title</ModalHeader>
-      <form
+      <Form
+        onSubmit={handleSubmit}
         id="edit-folder-title-form"
-        className={styles.editFolderTitle__form}
-        onSubmit={handleSubmit(onSubmit)}
+        className={styles.modal__content}
+        action={formAction}
+        method="POST"
       >
         <div>
-          <InputLabel htmlFor="folder-title-input">Title</InputLabel>
-          <Input
-            {...register('folderTitle', {
+          <TextField
+            {...register('title', {
               required: true,
-              minLength: 0,
-              maxLength: PROJECT_FOLDER_TITLE_MAX_LENGTH,
+              value: defaultTitle,
             })}
-            autoFocus
-            defaultValue={defaultTitle}
+            fieldState={getFieldState('title', formState)}
+            label="Title"
+            required
+            size="lg"
             id="folder-title-input"
             placeholder="Folder"
           />
         </div>
-      </form>
+      </Form>
       <ModalFooter
         before={
           <Button
             onClick={onClose}
-            disabled={isLoading}
+            disabled={isPending}
             variant="secondary-dimmed"
             size="md"
           >
@@ -110,14 +89,30 @@ const EditFolderTitleModal: React.FC<EditFolderTitleModalProps> = ({
             variant="primary"
             size="md"
             type="submit"
-            disabled={isLoading}
-            loading={isLoading}
+            disabled={isPending}
+            loading={isPending}
           >
             Save
           </Button>
         }
       />
-    </>
+    </Modal>
+  )
+}
+
+const EditFolderTitleModal: React.FC<EditFolderTitleModalProps> = ({
+  onClose,
+}) => {
+  const data = useProjectAwait()
+
+  return (
+    <Suspense>
+      <Await resolve={data?.project}>
+        {(resolvedProject) => (
+          <ModalContent onClose={onClose} project={resolvedProject} />
+        )}
+      </Await>
+    </Suspense>
   )
 }
 
