@@ -1,10 +1,7 @@
-'use client'
-import React, { useEffect, useId, useRef, useState } from 'react'
+import React, { useCallback, useEffect, useId, useRef, useState } from 'react'
 import deburr from 'lodash.deburr'
 import {
   exhaustivnessCheck,
-  FolderGetReq,
-  FolderGetRes,
   MAX_UPLOAD_FILE_SIZE,
   MULTIPART_UPLOAD_CHUNK_SIZE,
   TusHookPreFinishResponse,
@@ -14,194 +11,31 @@ import {
 } from '@valley/shared'
 import Uppy, { Meta, UppyFile } from '@uppy/core'
 import Tus from '@uppy/tus'
-import { TUSD_URL } from '../config/constants'
-import { getAuthTokens } from '../utils/accessToken'
-import { create } from 'zustand'
-import { immer } from 'zustand/middleware/immer'
 import { HttpRequest, HttpResponse } from 'tus-js-client'
 import { Folder, Project } from '@valley/db'
-import useSWR from 'swr'
-import { api } from '../api'
+import { useRevalidator, useRouteLoaderData } from '@remix-run/react'
+import { loader as rootLoader } from 'app/root'
+import { useUploadsStore } from 'app/stores/uploads'
+import { createUploadToken } from 'app/api/uploads'
+import { cache } from 'app/utils/client-cache'
+import { getFolderCacheKey } from 'app/routes/_user+/projects_.$projectId+/folder.$folderId'
+import { getProjectCacheKey } from 'app/routes/_user+/projects_.$projectId+/_layout'
 
-type Upload = {
-  id: string
-  normalizedName: string
-  filetype: string
-  bytesUploaded: number
-  totalBytes: number
-  progress: number
-  uploadError: string | null
-  isUploading: boolean
-  isUploaded: boolean
-}
-
-type UploadsState = {
-  uploads: Record<string, Upload>
-  uploadsCount: number
-  totalBytes: number
-  bytesUploaded: number
-  /** Bytes in second */
-  uploadSpeed: number
-  /** Unix timestamp of last execution of setUploadProgress */
-  lastUploadProgressTimestamp: number
-  lastBytesUploaded: number
-  /** Upload remaining time based on upload speed in seconds */
-  remainingTime: number
-  uploadSpeedIntervalID: NodeJS.Timeout | null
-  isUploading: boolean
-  folderId: Folder['id'] | null
-  projectId: Project['id'] | null
-}
-
-type UploadsAction = {
-  setIsUploading: (state: UploadsState['isUploading']) => void
-  setFileUploadError: (id: Upload['id'], err: Upload['uploadError']) => void
-  addUpload: (
-    upload: Pick<Upload, 'id' | 'totalBytes' | 'filetype' | 'normalizedName'>
-  ) => void
-  setFileUploadProgress: (
-    id: Upload['id'],
-    bytesUploaded: Upload['bytesUploaded']
-  ) => void
-  setFileUploaded: (id: Upload['id']) => void
-  clearSuccessfulUploads: () => void
-  setFolderId: (id: UploadsState['folderId']) => void
-  setProjectId: (id: UploadsState['projectId']) => void
-  setUploadSpeed: (uploadSpeed: UploadsState['uploadSpeed']) => void
-  updateUploadSpeed: () => void
-}
-
-export const useUploadsStore = create<UploadsState & UploadsAction>()(
-  immer((set) => ({
-    uploads: {},
-    uploadsCount: 0,
-    totalBytes: 0,
-    bytesUploaded: 0,
-    uploadSpeed: 0,
-    lastBytesUploaded: 0,
-    lastUploadProgressTimestamp: 0,
-    remainingTime: 0,
-    isUploading: false,
-    folderId: null,
-    projectId: null,
-    uploadSpeedIntervalID: null,
-    updateUploadSpeed: () =>
-      set((state) => {
-        state.uploadSpeed = state.bytesUploaded - state.lastBytesUploaded
-        state.lastBytesUploaded = state.bytesUploaded
-        state.remainingTime =
-          (state.totalBytes - state.bytesUploaded) / state.uploadSpeed
-      }),
-    setIsUploading: (newIsUploading) =>
-      set((state) => {
-        state.isUploading = newIsUploading
-      }),
-    setFolderId: (id) =>
-      set((state) => {
-        state.folderId = id
-      }),
-    setUploadSpeed: (uploadSpeed) =>
-      set((state) => {
-        state.uploadSpeed = uploadSpeed
-      }),
-    setProjectId: (id) =>
-      set((state) => {
-        state.projectId = id
-      }),
-    setFileUploadError: (id, err) =>
-      set((state) => {
-        if (!state.uploads[id]) {
-          throw new Error(
-            `Cannot set upload error to an unknown upload, got: ${id}`
-          )
-        }
-
-        state.uploads[id].isUploading = false
-        state.uploads[id].isUploaded = false
-        state.uploads[id].uploadError = err
-      }),
-    addUpload: (upload) =>
-      set((state) => {
-        state.uploads[upload.id] = {
-          ...upload,
-          bytesUploaded: 0,
-          isUploaded: false,
-          isUploading: true,
-          progress: 0,
-          uploadError: null,
-        }
-        state.uploadsCount += 1
-        state.totalBytes += upload.totalBytes
-      }),
-    setFileUploadProgress: (id, bytesUploaded) =>
-      set((state) => {
-        const upload = state.uploads[id]
-        if (!upload) {
-          throw new Error(
-            `Cannot set upload progress of an unknown upload, got: ${id}`
-          )
-        }
-
-        const prevBytesUploaded = upload.bytesUploaded
-        const progress = bytesUploaded / upload.totalBytes
-        state.uploads[id].uploadError = null
-        state.uploads[id].isUploading = true
-        state.uploads[id].bytesUploaded = bytesUploaded
-        state.uploads[id].progress = progress
-        state.uploads[id].isUploaded = false
-        state.bytesUploaded += bytesUploaded - prevBytesUploaded
-      }),
-    setFileUploaded: (id) =>
-      set((state) => {
-        if (!state.uploads[id]) {
-          throw new Error(
-            `Cannot set upload progress of an unknown upload, got: ${id}`
-          )
-        }
-
-        state.uploads[id].uploadError = null
-        state.uploads[id].isUploaded = true
-        state.uploads[id].isUploading = false
-      }),
-    clearSuccessfulUploads: () =>
-      set((state) => {
-        const erroredUploads: UploadsState['uploads'] = {}
-        let erroredUploadsCount = 0
-        let erroredUploadsTotalBytes = 0
-        let erroredUploadsBytesUploaded = 0
-
-        for (const id of Object.keys(state.uploads)) {
-          if (state.uploads[id].uploadError) {
-            erroredUploads[id] = state.uploads[id]
-            erroredUploadsCount += 1
-            erroredUploadsTotalBytes += state.uploads[id].totalBytes
-            erroredUploadsBytesUploaded += state.uploads[id].bytesUploaded
-          }
-        }
-
-        state.isUploading = false
-        state.uploads = erroredUploads
-        state.uploadsCount = erroredUploadsCount
-        state.totalBytes = erroredUploadsTotalBytes
-        state.bytesUploaded = erroredUploadsBytesUploaded
-      }),
-  }))
-)
+const isClientSide = typeof document !== 'undefined'
 
 type UseUploadProps = {
-  projectId: number
-  folderId: number
+  projectId: Project['id']
+  folderId: Folder['id']
 }
 
 export const useUpload = ({ projectId, folderId }: UseUploadProps) => {
-  const { mutate } = useSWR<FolderGetRes, FolderGetReq>(
-    '/projects/' + projectId + '/folders/' + folderId,
-    api({ isAccessTokenRequired: true })
-  )
-  const tokens = getAuthTokens()
   const inputId = useId()
+  const rootContext = useRouteLoaderData<typeof rootLoader>('root')
   const $root = useRef(null)
-  const $input = useRef<HTMLInputElement>(document.createElement('input'))
+  const $input = useRef<HTMLInputElement>(
+    isClientSide ? document.createElement('input') : null
+  )
+  const revalidator = useRevalidator()
   const setIsUploading = useUploadsStore((state) => state.setIsUploading)
   const setFolderId = useUploadsStore((state) => state.setFolderId)
   const setProjectId = useUploadsStore((state) => state.setProjectId)
@@ -219,32 +53,15 @@ export const useUpload = ({ projectId, folderId }: UseUploadProps) => {
   )
   const uploadSpeedIntervalID = useRef<NodeJS.Timeout>()
 
-  const getUppyHeaders = (): Record<string, string> => {
-    if (!tokens) return {}
-    return { Authorization: `Bearer ${tokens.accessToken}` }
-  }
+  const addUploadedFileToCache = async (file: TusHookPreFinishResponse) => {
+    await Promise.all([
+      cache.removeItem(getFolderCacheKey(file.folderId)),
+      cache.removeItem(getProjectCacheKey(file.projectId)),
+    ])
 
-  const addUploadedFileToSWRCache = (file: TusHookPreFinishResponse) => {
-    const newFile = {
-      bucket: file.bucket,
-      dateCreated: new Date(file.dateCreated),
-      exifMetadata: file.exifMetadata,
-      folderId: file.folderId,
-      id: file.id,
-      key: file.key,
-      name: file.name,
-      size: file.size,
-      thumbnailKey: file.thumbnailKey || null,
-      type: file.contentType,
-    }
-
-    mutate((data) => {
-      if (!data) return undefined
-      return {
-        ...data,
-        files: [...data.files, newFile],
-      }
-    })
+    try {
+      revalidator.revalidate()
+    } catch (e) {}
   }
 
   const handleUploadResponse = async (_req: HttpRequest, res: HttpResponse) => {
@@ -259,7 +76,7 @@ export const useUpload = ({ projectId, folderId }: UseUploadProps) => {
           case TusHookType.PRE_CREATE:
             break
           case TusHookType.PRE_FINISH:
-            addUploadedFileToSWRCache(parsedBody)
+            await addUploadedFileToCache(parsedBody)
             break
           default:
             exhaustivnessCheck(parsedBody)
@@ -282,55 +99,64 @@ export const useUpload = ({ projectId, folderId }: UseUploadProps) => {
       // Allows duplicate files
       onBeforeFileAdded: () => true,
     }).use(Tus, {
-      endpoint: TUSD_URL,
+      endpoint: rootContext?.ENV.TUSD_URL + '/files',
       chunkSize: MULTIPART_UPLOAD_CHUNK_SIZE,
-      headers: getUppyHeaders,
       onAfterResponse: handleUploadResponse,
     })
   )
 
-  const uploadFiles = async (fileList: FileList) => {
-    const files = [...fileList]
-    const fileIDs: string[] = []
-    let totalSize = 0
-    files.forEach((file) => (totalSize += file.size))
+  const uploadFiles = useCallback(
+    async (fileList: FileList) => {
+      const token = await createUploadToken({ projectId, folderId })
 
-    files.forEach(async (file) => {
-      const fileId = uppy.addFile(file)
-      const metadata: TusUploadMetadata = {
-        'normalized-name': deburr(file.name),
-        'upload-id': fileId,
-        'folder-id': folderId.toString(),
-        'project-id': projectId.toString(),
-        type: file.type,
-      }
-      uppy.setFileMeta(fileId, metadata)
-      fileIDs.push(fileId)
-      addUpload({
-        id: fileId,
-        totalBytes: file.size,
-        filetype: file.type,
-        normalizedName: metadata['normalized-name'],
+      const files = [...fileList]
+      const fileIDs: string[] = []
+      let totalSize = 0
+      files.forEach((file) => (totalSize += file.size))
+
+      files.forEach(async (file) => {
+        const fileId = uppy.addFile(file)
+        const metadata: TusUploadMetadata = {
+          'normalized-name': deburr(file.name),
+          'upload-id': fileId,
+          'user-id': token.userId,
+          'folder-id': token.uploadFolderId,
+          'project-id': token.uploadProjectId,
+          'upload-token': token.hash,
+          type: file.type,
+        }
+        uppy.setFileMeta(fileId, metadata)
+        fileIDs.push(fileId)
+        addUpload({
+          id: fileId,
+          totalBytes: file.size,
+          filetype: file.type,
+          normalizedName: metadata['normalized-name'],
+        })
       })
-    })
 
-    setIsUploading(true)
-    uploadSpeedIntervalID.current = setInterval(() => {
-      updateUploadSpeed()
-    }, 1000)
+      setIsUploading(true)
+      uploadSpeedIntervalID.current = setInterval(() => {
+        updateUploadSpeed()
+      }, 1000)
 
-    const res = await uppy.upload()
+      const res = await uppy.upload()
 
-    setIsUploading(false)
-    clearInterval(uploadSpeedIntervalID.current)
-    uppy.removeFiles(res?.successful?.map((e) => e.id) || [])
-  }
+      setIsUploading(false)
+      clearInterval(uploadSpeedIntervalID.current)
+      uppy.removeFiles(res?.successful?.map((e) => e.id) || [])
+    },
+    [addUpload, folderId, projectId, setIsUploading, updateUploadSpeed, uppy]
+  )
 
-  const handleUploadInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    if (!e.target.files) return
-    clearSuccessfulUploads()
-    uploadFiles(e.target.files)
-  }
+  const handleUploadInputChange = useCallback(
+    (e: React.ChangeEvent<HTMLInputElement>) => {
+      if (!e.target.files) return
+      clearSuccessfulUploads()
+      uploadFiles(e.target.files)
+    },
+    [clearSuccessfulUploads, uploadFiles]
+  )
 
   const register = <T extends HTMLElement = HTMLElement>(): {
     ref: React.RefObject<T>
@@ -384,7 +210,7 @@ export const useUpload = ({ projectId, folderId }: UseUploadProps) => {
       uppy.off('upload-error', handleUploadError)
       uppy.off('upload-success', handleUploadSuccess)
     }
-  }, [uppy])
+  }, [setFileUploadError, setFileUploadProgress, setFileUploaded, uppy])
 
   useEffect(() => {
     const handleChange = (e: Event) =>
@@ -392,7 +218,9 @@ export const useUpload = ({ projectId, folderId }: UseUploadProps) => {
         e as unknown as React.ChangeEvent<HTMLInputElement>
       )
 
-    document.body.appendChild($input.current)
+    if (!$input.current) return
+
+    isClientSide && document.body.appendChild($input.current)
     $input.current.multiple = true
     $input.current.hidden = true
     $input.current.type = 'file'
@@ -402,12 +230,21 @@ export const useUpload = ({ projectId, folderId }: UseUploadProps) => {
     setFolderId(folderId)
     setProjectId(projectId)
 
+    const inputRef = $input.current
+
     return () => {
-      $input.current.removeEventListener('change', handleChange)
-      $input.current.remove()
+      inputRef.removeEventListener('change', handleChange)
+      inputRef.remove()
     }
     // Should change `handleUploadInputChange` ref as hook props change
-  }, [folderId, projectId])
+  }, [
+    folderId,
+    handleUploadInputChange,
+    inputId,
+    projectId,
+    setFolderId,
+    setProjectId,
+  ])
 
   useEffect(() => {
     return () => {
