@@ -35,7 +35,7 @@ module.exports = {
       { name: 'Link', linkAttribute: 'to' },
       { name: 'NavLink', linkAttribute: 'to' },
     ],
-    'import/internal-regex': '^~/',
+    // 'import/internal-regex': '^~/',
     'import/resolver': {
       node: {
         extensions: ['.ts', '.tsx'],
