@@ -1,5 +1,5 @@
-export * from './Project'
-export * from './File'
-export * from './User'
-export * from './Folder'
-export * from './Tus'
+export * from './Project.js'
+export * from './File.js'
+export * from './User.js'
+export * from './Folder.js'
+export * from './Tus.js'
