@@ -1,4 +1,4 @@
-export * from './api'
-export * from './SerializedUser'
-export * from './SerializedProject'
-export * from './SerializedFolder'
+export * from './api/index.js'
+export * from './SerializedUser.js'
+export * from './SerializedProject.js'
+export * from './SerializedFolder.js'
