@@ -1,7 +1,9 @@
 import type { Meta, StoryObj } from '@storybook/react'
 import Button, { ButtonProps } from './Button'
-import { RefAttributes } from 'react'
 import { JSX } from 'react/jsx-runtime'
+import Divider from '../Divider/Divider'
+import Stack from '../Stack/Stack'
+import { ChevronRight, Plus } from 'geist-ui-icons'
 
 const meta: Meta<typeof Button> = {
   title: 'Components/Button',
@@ -10,6 +12,16 @@ const meta: Meta<typeof Button> = {
     layout: 'fullscreen',
   },
 }
+
+const variants: Array<ButtonProps['variant']> = [
+  'primary',
+  'secondary',
+  'danger',
+  'warning',
+  'tertiary',
+  'secondary-dimmed',
+  'tertiary-dimmed',
+]
 
 export default meta
 type Story = StoryObj<typeof Button>
@@ -20,15 +32,7 @@ export const Default: Story = {
   },
   argTypes: {
     variant: {
-      options: [
-        'primary',
-        'secondary',
-        'danger',
-        'warning',
-        'tertiary',
-        'secondary-dimmed',
-        'tertiary-dimmed',
-      ],
+      options: variants,
       control: { type: 'select' },
     },
     disabled: {
@@ -37,13 +41,66 @@ export const Default: Story = {
     loading: {
       control: { type: 'boolean' },
     },
+    before: {
+      control: { type: 'boolean' },
+    },
+    after: {
+      control: { type: 'boolean' },
+    },
+    fullWidth: {
+      control: { type: 'boolean' },
+    },
+    align: {
+      control: { type: 'select' },
+      table: {
+        defaultValue: { summary: 'center' },
+      },
+      options: ['start', 'center', 'end'],
+    },
     size: {
       options: ['sm', 'md', 'lg'],
       control: { type: 'select' },
     },
   },
-  render: (
-    args: JSX.IntrinsicAttributes &
-      (ButtonProps & RefAttributes<HTMLButtonElement>)
-  ) => <Button {...args} />,
+  render: ({
+    before,
+    after,
+    ...args
+  }: JSX.IntrinsicAttributes & ButtonProps) => (
+    <Stack gap={2} direction={'column'}>
+      <Button
+        {...args}
+        before={before ? <Plus /> : undefined}
+        after={after ? <ChevronRight /> : undefined}
+        size="sm"
+      />
+      <Button
+        {...args}
+        before={before ? <Plus /> : undefined}
+        after={after ? <ChevronRight /> : undefined}
+        size="md"
+      />
+      <Button
+        {...args}
+        before={before ? <Plus /> : undefined}
+        after={after ? <ChevronRight /> : undefined}
+        size="lg"
+      />
+      <Divider />
+      {variants.map((variant) => (
+        <Stack key={variant} gap={2} direction={'column'}>
+          <h2>{variant}</h2>
+          <Button variant={variant} size="sm">
+            {args.children || 'Label'}
+          </Button>
+          <Button variant={variant} size="md">
+            {args.children || 'Label'}
+          </Button>
+          <Button variant={variant} size="lg">
+            {args.children || 'Label'}
+          </Button>
+        </Stack>
+      ))}
+    </Stack>
+  ),
 }
