@@ -2,11 +2,12 @@ import cx from 'classnames'
 import React from 'react'
 import Paper, { PaperProps } from '../Paper/Paper'
 import styles from './Input.module.css'
+import { FormControlState, useFormControl } from '../FormControl'
 
-type InputProps = {
+export type InputProps = {
   before?: React.ReactNode
   after?: React.ReactNode
-  state?: 'default' | 'error' | 'valid'
+  state?: FormControlState['state']
   size?: 'md' | 'lg'
   paperProps?: Omit<PaperProps, 'asChild'>
 } & Omit<React.ComponentPropsWithRef<'input'>, 'size'>
@@ -18,19 +19,20 @@ const Input = React.forwardRef<HTMLInputElement, InputProps>(
       before,
       after,
       paperProps,
-      state = 'default',
+      state: propsState,
       size = 'md',
       ...props
     },
     ref
   ) {
+    const formControl = useFormControl()
+    const state = propsState || formControl?.state
+
     return (
       <Paper
         {...paperProps}
         variant="secondary"
         className={cx(styles.input, 'Input', paperProps?.className, {
-          [styles['input--withBefore']]: !!before,
-          [styles['input--withAfter']]: !!after,
           [styles['input--error']]: state === 'error',
           [styles['input--valid']]: state === 'valid',
           [styles['input--size-md']]: size === 'md',
