@@ -25,4 +25,4 @@ const Spinner: React.FC<
   )
 }
 
-export default Spinner
+export default React.memo(Spinner)
