@@ -5,21 +5,22 @@ import Paper, { PaperProps } from '../Paper/Paper'
 
 export type WrapperProps = PaperProps
 
-const Wrapper: React.FC<WrapperProps> = ({
-  children,
-  className,
-  variant = 'tertiary',
-  ...props
-}) => {
-  return (
-    <Paper
-      {...props}
-      variant={variant}
-      className={cx('Wrapper', styles.wrapper, className)}
-    >
-      {children}
-    </Paper>
-  )
-}
+const Wrapper = React.forwardRef<HTMLDivElement, WrapperProps>(
+  function WrapperWithRef(
+    { children, className, variant = 'tertiary', ...props },
+    ref
+  ) {
+    return (
+      <Paper
+        {...props}
+        ref={ref}
+        variant={variant}
+        className={cx('Wrapper', styles.wrapper, className)}
+      >
+        {children}
+      </Paper>
+    )
+  }
+)
 
 export default Wrapper
